more refactor + improving alignment
